Breakout Flags, RegularExpressions, and ReportLevel from SmartReader
