+++ src/SmartReader/ReportLevel.cs
@@ -1,0 +1,11 @@
+﻿namespace SmartReader
+{
+    /// <summary>The level of debug information to record</summary>
+    public enum ReportLevel
+    {
+        /// <summary>Only issues</summary>
+        Issue,
+        /// <summary>Every useful information</summary>
+        Info
+    }
+}
