+++ src/SmartReader/SmartReader.cs
@@ -14,51 +14,6 @@ using AngleSharp.Html.Parser;
 
 namespace SmartReader
 {
-    /// <summary>Flags that sets how aggressively remove potentially useless content</summary>
-    [Flags]
-    public enum Flags
-    {
-        /// <summary>Do not perform any cleaning</summary>
-        None = 0,
-        /// <summary>Remove unlikely content</summary>
-        StripUnlikelys = 1,
-        /// <summary>Remove content according that does not pass a certain threshold</summary>
-        WeightClasses = 2,
-        /// <summary>Clean content that does not look promising</summary>
-        CleanConditionally = 4
-    }
-
-    /// <summary>The level of debug information to record</summary>
-    public enum ReportLevel
-    {
-        /// <summary>Only issues</summary>
-        Issue,
-        /// <summary>Every useful information</summary>
-        Info
-    }
-
-    /// <summary>The different kinds regular expressions used to filter elements</summary>
-    public enum RegularExpressions
-    {
-        /// <summary>To remove elements unlikely to contain useful content</summary>
-        UnlikelyCandidates,
-        /// <summary>To find elements likely to contain useful content</summary>
-        PossibleCandidates,
-        /// <summary>Classes and tags that increases chances to keep the element</summary>
-        Positive,
-        /// <summary>Classes and tags that decreases chances to keep the element</summary>
-        Negative,
-        /// <summary>Extraneous elements</summary>
-        /// <remarks>Nota that this regular expression is not used anywhere at the moment</remarks>
-        Extraneous,
-        /// <summary>To detect byline</summary>
-        Byline,
-        /// <summary>To keep only useful videos</summary>
-        Videos,
-        /// <summary>To find sharing elements</summary>
-        ShareElements
-    }
-
     /// <summary>The main Reader class</summary>
     /// <remarks>
 	/// <para>This code is based on a port of the readability library of Firefox Reader View
